fix user

+++ server/src/main/java/ru/practicum/shareit/user/storage/DBUserStorageImpl.java
@@ -12,6 +12,7 @@ import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.repository.UserRepository;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,14 +25,25 @@ public class DBUserStorageImpl implements UserStorage {
 
     @Override
     public UserDto create(User user) throws ValidationException, ConflictException {
-        User user1 = userRepository.save(user);
-        UserDto userDto = userMapper.toUserDto(user1);
-        log.info("Пользователь успешно добавлен {}", user.getId());
-        return userDto;
+            User afterCheckUser = standardCheck(user);
+            Optional<User> userCheck = userRepository.findByEmail(afterCheckUser.getEmail());
+            if (userCheck.isPresent()) {
+                User afterSaveUser = userRepository.save(user);
+                userRepository.deleteById(afterSaveUser.getId());
+                throw new ConflictException("Пользователь уже есть");
+            }
+            User user1 = userRepository.save(afterCheckUser);
+            UserDto userDto = userMapper.toUserDto(user1);
+            log.info("Пользователь успешно добавлен {}", user.getId());
+            return userDto;
     }
 
     @Override
     public UserDto update(Long id, User user) throws ValidationException, ConflictException {
+        if (id <= 0) {
+            throw new ObjectNotFoundException("Пользователь не найден");
+        }
+
         User userBase = userRepository.findById(id).orElseThrow(() ->
                 new ObjectNotFoundException("Пользователь не найден"));
 
@@ -59,21 +71,48 @@ public class DBUserStorageImpl implements UserStorage {
 
     @Override
     public UserDto getUserDtoById(Long id) {
+        if (id <= 0) {
+            throw new ObjectNotFoundException("Пользователь не найден");
+        }
+
         return userMapper.toUserDto(userRepository.findById(id).orElseThrow(() ->
                 new ObjectNotFoundException("Пользователь не найден")));
     }
 
     @Override
     public User getUserById(Long id) {
+        if (id <= 0) {
+            throw new ObjectNotFoundException("Пользователь не найден");
+        }
+
         return userRepository.findById(id).orElseThrow(() ->
                 new ObjectNotFoundException("Пользователь не найден"));
     }
 
     @Override
     public boolean deleteUser(Long id) {
+        if (id <= 0) {
+            throw new ObjectNotFoundException("Пользователь не найден");
+        }
         User userBase = userRepository.findById(id).orElseThrow(() ->
                 new ObjectNotFoundException("Пользователь не найден"));
         userRepository.delete(userBase);
         return true;
+    }
+
+    private User standardCheck(User user) throws ValidationException, ConflictException {
+        if (user.getEmail() == null ||
+                user.getEmail().isEmpty() ||
+                user.getEmail().isBlank() ||
+                !user.getEmail().contains("@")) {
+            log.error("Неверно введен email: {}", user);
+            throw new ValidationException("Неверно введен email");
+        }
+
+        if (user.getName().isEmpty() || user.getEmail().isBlank() || user.getName().contains(" ")) {
+            log.error("Имя пользователя не может быть пустым: {}", user);
+            throw new ValidationException("Имя пользователя не может быть пустым");
+        }
+        return user;
     }
 }
